\name{AggregateReturns}
\alias{AggregateReturns}
\docType{data}

\title{Intra-day stock price returns data}

\description{This data set contains intra-day stock price returns (in per cent) of
  7 Indian banks with highest weight in the Bank NIFTY Index. The intra-day stock
  returns for 35 minutes before and after the event are used to define
  the event window. }

\usage{data(AggregateReturns)}

\author{Sargam Jain}

\keyword{datasets}
