\name{AMMData}
\alias{AMMData}
\docType{data}

\title{Data set containing firm returns, market returns, currency
  returns, and call money rate used for AMM estimation}

\description{This data set consists of daily time series for firm
  returns (Infosys and TCS), market returns (Nifty returns), currency
  returns (INR/USD), and call money rate. It is used to demonstrate
  augmented market model estimation.

  The data series is a daily time-series zoo object. The sample range for
  the data is from 2012-02-01 to 2014-01-31. All series are in per cent.
}

\usage{data(AMMData)}

\author{Vikram Bahure}

\keyword{AMMData}
