\name{AttackDates}
\alias{AttackDates}
\docType{data}

\title{Dates of terrorist attack}

\description{Accounting for different time zones across 33 global
  capital markets, this data set provides different dates for each
  capital market on the day of the terrorist attack of September 11, 2001
  EST. The data set is prepared using the information provided in 'The
  effects of terrorism on global capital markets' by Andrew H. Chen and
  Thomas F. Siems.}

\usage{data(AttackDates)}

\author{Sargam Jain}

\keyword{datasets}
