\name{EarningsDates}
\alias{EarningsDates}
\docType{data}

\title{Earnings announcement data}

\description{This data set contains earnings announcement records of
  companies belonging to the services and technology sectors between
  January 9, 2007 and February 28, 2007. The data set is prepared using
  the data provided in 'Event studies for financial research' by Doron
  Kliger and Gregory Gurevich.}

\usage{data(EarningsDates)}

\author{Sargam Jain}

\keyword{datasets}
