\name{eesData}

\docType{data}

\alias{eesData}

\title{It is the data-set used for event-study analysis.}

\description{
It is a time series object with daily series for S&P 500 and Nifty (NSE index).	
}

\usage{data(eesData)}

\format{An object with class attributes \code{zoo} containing resposne series for eventstudy.}

\examples{
    data(eesData)
}
\keyword{datasets}
