\name{eventslist}
\alias{eventslist}
\docType{data}
\title{
List of event date for a few companies.
}
\description{
A sample of event dates for few companies.
}
\usage{data(eventslist)}
\format{
  A data frame with 6 observations on the following 2 variables.
  \describe{
    \item{\code{unit}}{a character vector}
    \item{\code{when}}{a Date}
  }
}
\details{
This data set contains  event dates corresponding to a few test cases.
}
\source{
Home brewed test data.
}
\examples{
data(eventslist)
}
\keyword{datasets}
