\name{eventstudies-package}
\alias{eventstudies-package}
\docType{package}

\title{
  eventstudies: R package for conducting event studies
}

\description{
  \pkg{eventstudies} is an R package for conducting event studies and a
  platform for methodological research on event studies. The package
  supports several data adjustment procedures and inference procedures
  like bootstrap and the classical t-statistic.  is an R package for
  conducting event studies and a platform for methodological research on
  event studies.
}

\details{
  \tabular{ll}{
    Package: \tab eventstudies\cr
    Type: \tab Package\cr
    Version: \tab 1.2.1\cr
    Date: \tab 2019-01-09\cr
    License: \tab GPL 2\cr
    LazyLoad: \tab yes\cr
}

This package allows a dataset to be studied in an event-time frame and
perform parametric / non-parametric analysis using several inference
procedures.
}

\author{
Chirag Anand, Vikram Bahure, Vimal Balasubramaniam, Ajay Shah

Maintainer: Chirag Anand <anand.chirag@gmail.com>
}

\keyword{ eventstudies }

