\name{eventstudy-package}
\alias{eventstudy}
\docType{package}
\title{
Event study analysis.
}
\description{
This package is used to undertake event study analyses using R.
}
\details{
\tabular{ll}{
Package: \tab eventstudy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-20\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
This package allows a dataset to be converted into an event-time frame to undertake any further parametric / non-parametric analysis using various inference procedures.

}
\author{
Ajay Shah,Ashim Kapoor,Vimal Balasubramniam.

Maintainer: ajayshah@mayin.org
}
\keyword{ eventstudies }

