\name{eventstudyData}

\docType{data}

\alias{eventstudyData}

\title{It is the data-set used for event-study analysis.}

\description{
It is a time series object with daily series for S&P 500, Nifty and Net FII flows in India.	
}

\usage{data(eventstudyData)}

\format{An object with class attributes \code{zoo} containing resposne series for eventstudy.}

\seealso{
eventDays
}

\examples{
    data(eventstudyData)
}
\keyword{datasets}
