\name{firmExposuresData}
\alias{firmExposuresData}


\title{Data for testing the amm package and examples therein}

\description{This is a zoo data that is used for testing purposes}

\usage{data(firmExposuresData)}

\author{Vimal Balasubramaniam}

\examples{
library(zoo)
data(firmExposuresData)
str(test)
}

\keyword{firmExposuresData}