\name{IndexReturns}
\alias{IndexReturns}
\docType{data}

\title{Market indice returns data}

\description{This data set contains intra-day CNX NIFTY 50 index returns
  (in per cent) for days when RBI announced key interest rate
  cuts. These are used to estimate abnormal returns for stock price
  returns of 7 Indian banks with highest weight in BANK NIFTY Index.}

\usage{data(IndexReturns)}

\author{Sargam Jain}

\keyword{datasets}
