\name{inputdata}

\docType{data}

\alias{Identify extreme events data}

\title{Times series data}

\description{
  The time series data is available for S&P 500, VIX and Net flows of
  FII as a percentage of total market capitalisation for India. The
  sample range for the daily data is from 2000-02-10 to 2011-07-29.
}

\usage{inputdata}

\format{An object with class attributes \code{xts} and \code{zoo} containing three variables with 2036 observations.}

\keyword{datasets}
