\name{inr}
\alias{inr}
\docType{data}
\title{
Exchange rate for Indian rupee: INR/USD
}
\description{
A sample of INR/USD rates from 1990 to 2011. It is stored as an xts object.
}
\usage{data(inr)}
\format{ It is an xts object containing daily data from 1990 to 2011.  The format is: chr 'inr.rda'
}
\examples{
data(inr)
}
\keyword{datasets}
