\name{inr.rda}
\alias{inr.rda}
\docType{data}
\title{
INR/USD
}
\description{
A sample of INR/USD rates from 1990 to 2011. It is stored as an xts object.  
}
\usage{data(inr.rda)}
\format{
  The format is:
 chr "inr.rda"
}
\examples{
data(inr.rda)
## maybe str(inr.rda) ; plot(inr.rda) ...
}
\keyword{datasets}
