\name{KGEarningsDates}
\alias{KGEarningsDates}
\docType{data}

\title{Earnings announcement data for replication of a standard event study}

\description{This data set contains earnings announcement records of
  companies belonging to the services and technology sectors between
  January 9, 2007 and February 28, 2007.}

\usage{data(KGEarningsDates)}

\references{
  \cite{Doron Kliger and Gregory Gurevich (2014).
    Event studies for financial research -- A comprehensive guide.
  }
}

\author{Sargam Jain}

\keyword{datasets}
