\name{KGMarketReturns}
\alias{KGMarketReturns}
\docType{data}

\title{Market indice returns data}

\description{This data set contains S&P 500 Index returns (in per cent)
  between January 9, 2007 and February 28, 2007. These are used to
  estimate abnormal returns for stock price returns of 670 individual
  firms in service and technology sector.}

\usage{data(KGMarketReturns)}

\references{
  \cite{Doron Kliger and Gregory Gurevich (2014).
    Event studies for financial research -- A comprehensive guide.
  }
}

\author{Sargam Jain}

\keyword{datasets}
