\name{KGStockReturns}
\alias{KGStockReturns}
\docType{data}

\title{Stock price returns data}

\description{This data set contains stock price returns (in per cent)
  of 670 individual firms between January 9, 2007 and February 28,
  2007 in service and technology sector.}

\usage{data(KGStockReturns)}

\references{
  \cite{Doron Kliger and Gregory Gurevich (2014).
    Event studies for financial research -- A comprehensive guide.
  }
}

\author{Sargam Jain}

\keyword{datasets}
