\name{KGSurpriseCategory}
\alias{KGSurpriseCategory}
\docType{data}

\title{Classification of stocks}

\description{This data set classifies the 670 firms analysed into 3
  categories: good, bad, and medium -- on the basis of unexpected
  deviation of realized earnings from the forecast following the
  earnings' announcements.}

\usage{data(KGSurpriseCategory)}

\references{
  \cite{Doron Kliger and Gregory Gurevich (2014).
    Event studies for financial research -- A comprehensive guide.
  }
}

\author{Sargam Jain}

\keyword{datasets}
