\name{nifty.index}
\alias{nifty.index}
\docType{data}
\title{
NSE Nifty index from 2004 to 2012
}
\description{
A sample of Nifty index return from 1990 to 2012. It is stored as an tim series object object.
}
\usage{data(nifty.index)}
\format{ It is an time series object containing daily data from 2003 to 2012.  The format is: chr 'nifty.index.rda'
}
\examples{
data(nifty.index)
}
\keyword{datasets}
