\name{numbers2words}
\alias{numbers2words}

\title{
  Converts numbers to words
}

\description{
  Converts numbers to words
 }

\usage{
numbers2words(x)
}

\arguments{
  \item{x}{'x' is number which is to be converted into words}
}

\value{
  Output is character string
}


\examples{
output <- numbers2words(1234)
}