\name{OtherReturns}
\alias{OtherReturns}
\docType{data}

\title{Data set containing daily returns of Nifty index, USD INR, call momey
  rate, and S&P 500 index}

\description{This data set consists of daily time series of market
  returns (Nifty index and S&P 500 index), currency returns (USD/INR),
  and call money rate.

  The data series is a daily time-series zoo object. All series are in
  per cent.
}

\usage{data(OtherReturns)}

\author{Chirag Anand}

\keyword{OtherReturns}
