\name{RateCuts}
\alias{RateCuts}
\docType{data}

\title{Dates of RBI key interest rate cuts}

\description{This data set contains dates of announcements of key
  interest rate cuts by Reserve Bank of India (RBI). For each
  announcement date, each company is considered a different
  entity. For example, for two different rate cuts say, on April 17,
  2012 and January 29, 2013, ICICI Bank is considered a separate
  entity.}

\usage{data(RateCuts)}

\author{Sargam Jain}

\keyword{datasets}
