\name{sp500}

\docType{data}

\alias{sp500}

\title{Times series data}

\description{
  It is time series data for S&P 500 from 2000-02-10 to 2011-07-29. 
}

\usage{data(sp500)}

\format{An object with class attributes \code{xts} and \code{zoo} containing a variable with 2036 observations.}

\examples{
    data(sp500)
}
\keyword{datasets}
