\name{SplitDates}

\docType{data}

\alias{SplitDates}

\title{It is the data-set used for event-study analysis.}

\description{
It has stock split event dates for the BSE index firms with two columns 'unit' which has firm names and 'when' which has event date for the firm.
}

\usage{data(SplitDates)}

\format{An object with class attributes \code{zoo} containing resposne series for eventstudy and a data frame with stock split event dates.}

\examples{
    data(SplitDates)
}
\keyword{datasets}
