\name{StockIndiceReturns}
\alias{StockIndiceReturns}
\docType{data}

\title{Stock indice returns data}

\description{This data set contains daily stock indice returns (in per cent) of
  33 global capital markets around September 11, 2009. The daily stock
  indice returns for 30 days pre and post the terrorist attack are used
  in the event studies analysis. The data set is prepared using
  the information provided in 'The effects of terrorism on global
  capital markets' by Andrew H. Chen and Thomas F. Siems.}

\usage{data(StockIndiceReturns)}

\author{Sargam Jain}

\keyword{datasets}
