\name{StockPriceReturns}

\docType{data}

\alias{StockPriceReturns}

\title{An example dataset of stock price returns to perform eventstudy analysis.}

\description{Daily stock price returns for the index constituents of the Bombay Stock Exchange Index (SENSEX) and the National Stock Exchange of India Index (NIFTY), NIFTY index, and the Rupee-Dollar exchange rate returns.}


\usage{data(StockPriceReturns)}

\format{\code{zoo}}

\examples{
    data(StockPriceReturns)
}
\keyword{datasets}
