\name{StockReturns}
\alias{StockReturns}
\docType{data}

\title{Stock price returns data}

\description{This data set contains stock price returns (in per cent)
  of 670 individual firms between January 9, 2007 and February 28,
  2007 in service and technology sector. The data set is prepared using
  the data provided in 'Event studies for financial research' by Doron
  Kliger and Gregory Gurevich.}

\usage{data(StockReturns)}

\author{Sargam Jain}

\keyword{datasets}
