\name{SurpriseCategory}
\alias{SurpriseCategory}
\docType{data}

\title{Classification of stocks}

\description{This data set classifies the 670 firms analysed into 3
  categories: good, bad, and medium -- on the basis of unexpected
  deviation of realized earnings from the forecast following the
  earnings' announcements. The data set is prepared using
  the data provided in 'Event studies for financial research' by Doron
  Kliger and Gregory Gurevich.}

\usage{data(SurpriseCategory)}

\author{Sargam Jain}

\keyword{datasets}
