\name{TerrorIndiceReturns}
\alias{TerrorIndiceReturns}
\docType{data}

\title{Stock indice returns data}

\description{This data set contains daily stock indice returns (in per cent) of
  33 global capital markets around September 11, 2009. The daily stock
  indice returns for 30 days pre and post the terrorist attack are used
  in the event studies analysis.}

\usage{data(TerrorIndiceReturns)}

\references{
  \cite{Andrew H Chen and Thomas F Siems (2004).
    The effects of terrorism on global capital markets.
    European Journal of Political Economy, 20(1), 349-366.
    \url{http://dev.wcfia.harvard.edu/sites/default/files/ChenSiems2004.pdf}
  }
}

\author{Sargam Jain}

\keyword{datasets}
