\name{EESData}
\alias{EESData}
\docType{data}

\title{Returns data used for extreme events analysis}

\description{This data set is used to demonstrate extreme events study
  functionality of the package. It contains daily returns data (in per
  cent) of S&P 500 and the NIFTY Index.}

\usage{data(EESData)}

\author{Chirag Anand}

\keyword{datasets}
