\name{eventstudies-package}
\alias{eventstudies-package}
\docType{package}

\title{
  eventstudies: R package for conducting event studies
}

\description{
  \pkg{eventstudies} is an R package for conducting event studies and a
  platform for methodological research on event studies.
}

\details{
  \tabular{ll}{
    Package: \tab eventstudies\cr
    Type: \tab Package\cr
    Version: \tab 1.1\cr
    Date: \tab 2014-03-28\cr
    License: \tab GPL 2\cr
    LazyLoad: \tab yes\cr
}

This package allows a dataset to be studied in an event-time frame and
perform parametric / non-parametric analysis using several inference
procedures.
}

\author{
Chirag Anand, Vikram Bahure, Vimal Balasubramaniam, Ajay Shah

Maintainer: Vikram Bahure <economics.vikram@gmail.com>
}

\keyword{ eventstudies }

