\name{eventstudy-package}
\alias{eventstudy}
\docType{package}
\title{
Event study analysis.
}
\description{
The is a package to do Event Studies using R.
}
\details{
\tabular{ll}{
Package: \tab eventstudy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-20\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
The 2 most important functions of this package are phys2eventtime and inference.Ecar.
The first of these converts the index of the zoo object containing the events
to event time while the second one performs bootstrap inference for the main 
graph of the event study.

}
\author{
Ajay Shah,Ashim Kapoor,Vimal Balasubramniam.

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
}
\keyword{ eventstudies }
\seealso{
}
\examples{
}
