\name{eventstudies-package}
\alias{eventstudies}
\docType{package}

\title{
Event study analysis.
}

\description{
This package is used to undertake event study analysis using R.
}

\details{
\tabular{ll}{
Package: \tab eventstudies\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-07-16\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}

This package allows a dataset to be studied in an event-time frame and
perform parametric / non-parametric analysis using several inference
procedures.
}

\author{
Ajay Shah, Chirag Anand, Vikram Bahure, Vimal Balasubramaniam

Maintainer: Vikram Bahure <economics.vikram@gmail.com>
}

\keyword{ eventstudies }

