\name{eventstudy-package}
\alias{eventstudy}
\docType{package}

\title{
Event study analysis.
}

\description{
This package is used to undertake event study analyses using R.
}

\details{
\tabular{ll}{
Package: \tab eventstudy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-04\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}

This package allows a dataset to be studied in an event-time frame and
perform parametric / non-parametric analysis using several inference
procedures.
}

\author{
Ajay Shah,Vimal Balasubramniam, Vikram Bahure

Maintainer: vimsaa@gmail.com
}

\keyword{ eventstudies }

