\name{sp500}

\docType{data}

\alias{Identify extreme events data}

\title{Times series data}

\description{
  It is time series data for S&P 500 from 2000-02-10 to 2011-07-29. 
}

\usage{sp500}

\format{An object with class attributes \code{xts} and \code{zoo} containing three variables with 2036 observations.}

\keyword{datasets}
