\name{SplitDates}
\alias{SplitDates}
\docType{data}

\title{Data set of events used to perform event study analysis}

\description{This data set contains stock split event dates for the index
  constituents of the Bombay Stock Exchange index (SENSEX). The data
  format follows the required format in the function
  \code{phys2eventtime}, with two columns 'outcome.unit' (firm name) and
  'event.when' (stock split date).}

\usage{data(SplitDates)}

\author{Vikram Bahure}

\keyword{datasets}
