\name{StockPriceReturns}

\docType{data}

\alias{StockPriceReturns}

\title{Sample data containing stock price returns}

\description{This data set contains stock price returns of 30 major
  stocks on the National Stock Exchange (NSE) of India for a period of 23
  years.}

\usage{data(StockPriceReturns)}

\format{\code{zoo}}

\examples{
    data(StockPriceReturns)
}
\keyword{datasets}
