\name{StockPriceReturns}

\docType{data}

\alias{StockPriceReturns}

\title{An example dataset of stock price returns to perform eventstudy analysis.}

% FIXME: Remove NAs in the dataset? If we do complete.cases then all rows will be removed
\description{This data set contains stock price returns of 30 major
  stocks on the National Stock Exchange (NSE) of India for a period of 23
  years.}

\usage{data(StockPriceReturns)}

\format{\code{zoo}}

\examples{
    data(StockPriceReturns)
}
\keyword{datasets}
