\name{StockPriceReturns}
\alias{StockPriceReturns}
\docType{data}

\title{Stock price returns data}

\description{This data set contains stock price returns (in per cent) of
  30 major stocks on the National Stock Exchange (NSE) of India.}

\usage{data(StockPriceReturns)}

\author{Vikram Bahure}

\keyword{datasets}
