\name{StockPriceReturns}

\docType{data}

\alias{StockPriceReturns}

\title{It is the data-set used for event-study analysis.}

\description{
It has stock price returns for index constituents of Bombay Stock Exchange (BSE).
}

\usage{data(StockPriceReturns)}

\format{An object with class attributes \code{zoo} containing resposne series for eventstudy.}

\examples{
    data(StockPriceReturns)
}
\keyword{datasets}
