\name{TerrorAttack}
\alias{TerrorAttack}
\docType{data}

\title{Dates of terrorist attack}

\description{Accounting for different time zones across 33 global
  capital markets, this data set provides different dates for each
  capital market on the day of the terrorist attack of September 11, 2001
  EST.}

\usage{data(TerrorAttack)}

\references{
  \cite{Andrew H Chen and Thomas F Siems (2004).
    The effects of terrorism on global capital markets.
    European Journal of Political Economy, 20(1), 349-366.
    \url{http://dev.wcfia.harvard.edu/sites/default/files/ChenSiems2004.pdf}
  }
}

\author{Sargam Jain}

\keyword{datasets}
